pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "google-genai>=0.3.0",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    "requests>=2.32.4",
    "starlette>=0.47.2",
]
